web: reorganize hooks and helpers
