+++ web/src/components/storage/VolumeGroupEditor.tsx
@@ -27,8 +27,7 @@ import { STORAGE as PATHS } from "~/routes/paths";
 import { apiModel } from "~/api/storage/types";
 import { model } from "~/types/storage";
 import { contentDescription } from "~/components/storage/utils/volume-group";
-import { useVolumeGroup } from "~/hooks/storage/model";
-import useDeleteVolumeGroup from "~/hooks/storage/delete-volume-group";
+import { useVolumeGroup, useDeleteVolumeGroup } from "~/hooks/storage/volume-group";
 import DeviceMenu from "~/components/storage/DeviceMenu";
 import DeviceHeader from "~/components/storage/DeviceHeader";
 import MountPathMenuItem from "~/components/storage/MountPathMenuItem";
