--- web/src/hooks/storage/update-api-model.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) [2025] SUSE LLC
- *
- * All Rights Reserved.
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation; either version 2 of the License, or (at your option)
- * any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, contact SUSE LLC.
- *
- * To contact SUSE LLC about this file by physical or electronic mail, you may
- * find current contact information at www.suse.com.
- */
-
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { setConfigModel } from "~/api/storage";
-import { apiModel } from "~/api/storage/types";
-
-type UpdateApiModelFn = (apiModel: apiModel.Config) => void;
-
-function useUpdateApiModel(): UpdateApiModelFn {
-  const queryClient = useQueryClient();
-  const query = {
-    mutationFn: (apiModel: apiModel.Config) => setConfigModel(apiModel),
-    onSuccess: () => queryClient.invalidateQueries({ queryKey: ["storage"] }),
-  };
-
-  const { mutate } = useMutation(query);
-  return mutate;
-}
-
-export { useUpdateApiModel as default };
-export type { UpdateApiModelFn };
